% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{posterior_set}
\alias{posterior_set}
\title{posterior_set}
\format{
An object of class \code{data.frame} with 1000 rows and 22 columns.
}
\usage{
posterior_set
}
\description{
posterior_set
}
\keyword{datasets}
