% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sex_behave_data}
\alias{sex_behave_data}
\title{sex_behave_data}
\format{
An object of class \code{list} of length 9.
}
\usage{
sex_behave_data
}
\description{
sex_behave_data
}
\keyword{datasets}
