% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{transmission_model}
\alias{transmission_model}
\title{Transmission model at a time step}
\usage{
transmission_model(t, initials, par)
}
\arguments{
\item{t}{time}

\item{initials}{epidemic states at each time step (population distribution in each of S, I, T, and P states)}

\item{par}{a list of parameter values}
}
\description{
Transmission model at a time step
}
